Added a method to the TcpProxy extracted class from TcpClientSession
