+++ src/main/java/com/github/steveice10/packetlib/tcp/TcpProxy.java
@@ -1,10 +1,16 @@
 package com.github.steveice10.packetlib.tcp;
 
 import com.github.steveice10.packetlib.ProxyInfo;
+import io.netty.channel.ChannelHandlerContext;
+import io.netty.channel.ChannelInboundHandlerAdapter;
 import io.netty.channel.ChannelPipeline;
+import io.netty.handler.codec.haproxy.*;
 import io.netty.handler.proxy.HttpProxyHandler;
 import io.netty.handler.proxy.Socks4ProxyHandler;
 import io.netty.handler.proxy.Socks5ProxyHandler;
+
+import java.net.Inet4Address;
+import java.net.InetSocketAddress;
 
 public class TcpProxy {
 
@@ -47,5 +53,23 @@ public class TcpProxy {
         }
     }
 
+    public void addHAProxySupport(ChannelPipeline pipeline, InetSocketAddress clientAddress) {
+            pipeline.addFirst("proxy-protocol-packet-sender", new ChannelInboundHandlerAdapter() {
+                @Override
+                public void channelActive(ChannelHandlerContext ctx) throws Exception {
+                    HAProxyProxiedProtocol proxiedProtocol = clientAddress.getAddress() instanceof Inet4Address ? HAProxyProxiedProtocol.TCP4 : HAProxyProxiedProtocol.TCP6;
+                    InetSocketAddress remoteAddress = (InetSocketAddress) ctx.channel().remoteAddress();
+                    ctx.channel().writeAndFlush(new HAProxyMessage(
+                            HAProxyProtocolVersion.V2, HAProxyCommand.PROXY, proxiedProtocol,
+                            clientAddress.getAddress().getHostAddress(), remoteAddress.getAddress().getHostAddress(),
+                            clientAddress.getPort(), remoteAddress.getPort()
+                    ));
+                    ctx.pipeline().remove(this);
+                    ctx.pipeline().remove("proxy-protocol-encoder");
+                    super.channelActive(ctx);
+                }
+            });
+            pipeline.addFirst("proxy-protocol-encoder", HAProxyMessageEncoder.INSTANCE);
+        }
 
 }
